Move all user related servlets into rest/admin/users.py
